refactor(completion): split import map extraction (#8846)
